Shift pytest fixtures to conftest.py
